requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.72.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==6.2.5
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.2.0
unicorn==1.0.3
urllib3==1.26.9
uvicorn==0.17.0
